Feat: Implement CRUD operations for idioms and add search request/response schemas
